Move some stuff to separate PermutedIndices module
